# src/main/resources/api.yml
...
swagger: '2.0'
...
info:
  version: "0.0.1"
  title: Business Process API

# Describe your paths here
paths:
...
  /group:
    get:
      operationId: getProcessInstanceGroups
      parameters:
      - in: query
        name: partyID
        type: string
        description: Identifier of the party
      - in: query
        name: relatedProducts
        type: array
        description: Related products
      - in: query
        name: relatedProductCategories
        type: array
        description: Related product categories
      - in: query
        name: tradingPartnerIDs
        type: array
        description: Identifier of the corresponsing trading partner ID
      - in: query
        name: initiationDateRange
        type: string
        description: Initiation date range for the first process instance in the group
      - in: query
        name: lastActivityDateRange
        type: string
        description: Last activity date range. It is the latest submission date of the document to last process instance in the group
      - in: query
        name: offset
        type: integer
        description: Offset of the first result among the complete result set satisfying the given criteria
        default: 0
      - in: query
        name: limit
        type: integer
        description: Number of results to be included in the result set
        default: 10
      - in: query
        name: archived
        type: boolean
        default: false
      - in: query
        name: collaborationRole
        type: string
        required: false
      - name: Authorization
        in: header
        type: string
        required: true
      produces:
      - application/json
      responses:
        200:
          description: Retrieved the groups successfully
          schema:
            $ref: '#/definitions/ProcessInstanceGroupResponse'
    post:
      operationId: saveProcessInstanceGroup
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: processInstanceGroup
        description: The content of the process instance group to be saved
        required: true
        schema:
          $ref: '#/definitions/ProcessInstanceGroup'
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        201:
          description: Saved the group successfully
  /group/filters:
    get:
      operationId: getProcessInstanceGroupFilters
      parameters:
      - in: query
        name: partyID
        type: string
        description: Identifier of the party
      - in: query
        name: relatedProducts
        type: array
        description: Related products
      - in: query
        name: relatedProductCategories
        type: array
        description: Related product categories
      - in: query
        name: tradingPartnerIDs
        type: array
        description: Identifier of the corresponsing trading partner ID
      - in: query
        name: initiationDateRange
        type: string
        description: Initiation date range for the first process instance in the group
      - in: query
        name: lastActivityDateRange
        type: string
        description: Last activity date range. It is the latest submission date of the document to last process instance in the group
      - in: query
        name: archived
        type: boolean
        default: false
      - in: query
        name: collaborationRole
        type: string
        required: false
      - name: Authorization
        in: header
        type: string
        required: true
      produces:
      - application/json
      responses:
        200:
          description: Retrieved the filters successfully
          schema:
            $ref: '#/definitions/ProcessInstanceGroupFilter'
  /group/{ID}:
    get:
      operationId: getProcessInstanceGroup
      parameters:
      - in: path
        name: ID
        type: string
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        200:
          description: Retrieved the groups successfully
          schema:
            $ref: '#/definitions/ProcessInstanceGroup'
    delete:
      operationId: deleteProcessInstanceGroup
      parameters:
      - in: path
        name: ID
        type: string
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        200:
          description: Deleted the groups successfully
  /group/{ID}/archive:
    post:
      operationId: archiveGroup
      parameters:
      - in: path
        name: ID
        type: string
        description: Identifier of the process instance group to be archived
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        200:
          description: Archived the group successfully
          schema:
            $ref: '#/definitions/ProcessInstanceGroup'
  /group/{ID}/restore:
    post:
      operationId: restoreGroup
      parameters:
      - in: path
        name: ID
        type: string
        description: Identifier of the process instance group to be restored
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        200:
          description: Restored the group successfully
          schema:
            $ref: '#/definitions/ProcessInstanceGroup'
  /group/archive-all:
    post:
      operationId: archiveAllGroups
      parameters:
      - in: query
        name: partyID
        type: string
        description: Identifier of the party of which groups will be archived
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        200:
          description: Archived the groups successfully
  /group/delete-all:
    post:
      operationId: deleteAllArchivedGroups
      parameters:
      - in: query
        name: partyID
        type: string
        description: Identifier of the party of which groups will be deleted
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        200:
          description: Deleted the groups successfully
  /group/{ID}/process-instance:
    get:
      operationId: getProcessInstancesOfGroup
      produces:
      - application/json
      parameters:
      - in: path
        name: ID
        type: string
        description: Identifier of the process instance group for which the associated process instances will be retrieved
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        200:
          description: Retrieved the process instances successfully
          schema:
            type: array
            items:
              $ref: '#/definitions/ProcessInstance'
    post:
      operationId: addProcessInstanceToGroup
      parameters:
      - in: path
        name: ID
        type: string
        description: Identifier of the process instance group to which a new process instance id is added
        required: true
      - in: query
        name: processInstanceID
        type: string
        description: Identifier of the process instance to be added
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        200:
          description: Added the process instance to the group successfully
          schema:
            $ref: '#/definitions/ProcessInstanceGroup'
    delete:
      operationId: deleteProcessInstanceFromGroup
      parameters:
      - in: path
        name: ID
        type: string
        description: Identifier of the process instance group from which the process instance id is deleted
        required: true
      - in: query
        name: processInstanceID
        type: string
        description: Identifier of the process instance to be deleted
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        200:
          description: Deleted the process instance from the group successfully
          schema:
            $ref: '#/definitions/ProcessInstanceGroup'
definitions:
...
  ProcessInstance:
    title: Business Process Instance
    type: object
    properties:
      processInstanceID:
        type: string
      creationDate:
        type: string
      processID:
        type: string
      status:
        type: string
        enum:
        - STARTED
        - WAITING
        - ABORTED
        - COMPLETED
  ProcessInstanceGroup:
    title: Process Instance Group Object
    type: object
    properties:
      ID:
        type: string
      partyID:
        type: string
      processInstanceIDs:
        type: array
        items:
          type: string
      archived:
        type: boolean
      collaborationRole:
        type: string
      associatedGroups:
        type: array
        items:
          type: string
      firstActivityTime:
        type: string
      lastActivityTime:
        type: string
  ProcessInstanceGroupResponse:
    title: Process Instance Group Response
    type: object
    properties:
      processInstanceGroups:
        type: array
        items:
          $ref: '#/definitions/ProcessInstanceGroup'
      size:
        type: integer
  ProcessInstanceGroupFilter:
    title: Process Instance Group Filters
    description: Contains the filtering criteria for Process Instance Group objects
    type: object
    properties:
      startDate:
        type: string
      endDate:
        type: string
      tradingPartnerIDs:
        type: array
        items:
          type: string
      tradingPartnerNames:
        type: array
        items:
          type: string
      relatedProducts:
        type: array
        items:
          type: string
      relatedProductCategories:
        type: array
        items:
          type: string
...